John conroy/cleanup cells UI (#2137)
* Remove demo cells components

* Delete unused cells components

* Move cells components into components dir

* Rename cells page

* Add changelog
