--- CHANGELOG-vertical-stacked-bar-chart
@@ -1,1 +1,0 @@
-- Add a reusable vertical stacked bar chart to shared styles.
