+++ context/app/static/js/components/cells/AutocompleteEntity/index.js
@@ -1,0 +1,3 @@
+import AutocompleteEntity from './AutocompleteEntity';
+
+export default AutocompleteEntity;
