+++ context/app/static/js/components/cells/CellsService.js
