+++ context/app/static/js/components/cells/DatasetsTable/index.js
@@ -1,0 +1,3 @@
+import DatasetsTable from './DatasetsTable';
+
+export default DatasetsTable;
