+++ context/app/static/js/pages/Cells/index.js
@@ -1,0 +1,3 @@
+import Cells from './Cells';
+
+export default Cells;
