--- context/app/static/js/pages/CellsAPIDemo/index.js
@@ -1,3 +1,0 @@
-import CellsAPIDemo from './CellsAPIDemo';
-
-export default CellsAPIDemo;
